requirements.txt
Django >= 3.2.4,<3.3
djangorestframework >=3.12.4,<3.13
#adding postgres django connector
psycopg2>=2.8.6,<2.9
#documentation
drf-spectacular>=0.15,<0.16

pillow>=8.2.0,<8.3.0

django-cors-headers

# added myself
python-decouple
